6 add core game functionality (#7)
Added the following:
- Playable map with 2 layouts (Orbital)
- 2 MCOMs per sector
- Score functionality for the MCOMs
- MCOMs keep their remaining time when diffused
- Overhauled the Map initialisation to make it easier to set up
- Added pre round timer setup with UI and sounds
- When both MCOMs are destroyed all players are now undeployed and wait
for the next round
- Debug MCOMs are added for AI to play the objective (can be toggled)
- Debug mode to play as defence has been added
- A manual player alive check is in place as Undeploying the player via
the editor is not updating the default game alive state causing issues
with the boundaries and UI conflicting states.
- Game ends after round 2

The current layout is inspired by the proposed images by Deadzigs.
Alternations were made to round 1 due to visibility and default game
boundaries. However this build is mainly to test the core functionality.

The attacking team score has not been implemented yet, this is planned
to be done later. Focus for now is testing the